Adjusting modal funcionalities and dashboard form

+++ src/app/views/dashboard/dashboard-reader/dashboard-reader.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, ViewChild } from '@angular/core';
+import { Component, Inject, OnInit, ViewChild } from '@angular/core';
 import { BarecodeScannerLivestreamComponent } from "ngx-barcode-scanner";
 import { Observable, of, Subject } from "rxjs";
 import { debounceTime, distinctUntilChanged, switchMap } from "rxjs/operators";
@@ -6,6 +6,7 @@ import Swal from 'sweetalert2';
 import * as _ from 'lodash';
 import { ActivatedRoute } from "@angular/router";
 import { LocalStorageService } from "../../../core/services";
+import { MAT_DIALOG_DATA } from "@angular/material/dialog";
 
 @Component({
   selector: 'app-dashboard-reader',
@@ -27,7 +28,8 @@ export class DashboardReaderComponent implements OnInit {
 
   constructor(
     private route: ActivatedRoute,
-    private localStorage: LocalStorageService
+    private localStorage: LocalStorageService,
+    @Inject(MAT_DIALOG_DATA) public data: any
   ) { }
 
   ngOnInit() {
@@ -64,18 +66,35 @@ export class DashboardReaderComponent implements OnInit {
     this.protocolReader = JSON.parse(protocol);
     let user = localStorage.getItem('appUser');
     user = JSON.parse(user);
-    let group_id = this.route.snapshot.paramMap.get('group_id');
 
-    let groups = _.filter(this.groupsReader, {'id': 1});
+    let groups = _.filter(this.groupsReader, {'id': parseInt(this.data.group_id)});
     let participants = _.filter(groups[0].participants, {'registration_code': parseInt(code)});
 
+    let new_protocol = {
+      'id': participants[0].id,
+      'moderator_id': user['id'],
+      'group_reader_id': parseInt(this.data.group_id),
+      'participant_name': participants[0].name,
+      'registration_code': participants[0].registration_code,
+      'type': groups[0].type,
+      'date_reader': '2019-09-24 00:00:00'
+    };
+
+    let participantsExist = _.filter(this.protocolReader, {'registration_code': parseInt(code)});
+
     if (participants.length > 0) {
+      if (!_.isEmpty(participantsExist)) {
+        Swal.fire('Ops!', 'Candidato já foi escaneado!', 'error');
+        return of('O candidato já foi escaneado!');
+      }
+
       Swal.fire('Candidato: ' + participants[0].name,
         'codigo inscricao: ' + participants[0].registration_code,
       );
-      this.protocolReader.push(participants[0]);
+
+      this.protocolReader.push(new_protocol);
       this.localStorage.setItem('protocols', JSON.stringify(this.protocolReader));
-      return of('Participante encontrado!' + participants[0].name);
+      return of('Participante encontrado: ' + participants[0].name);
     }
 
     return of('O codigo de inscrição está invalido!');
@@ -83,6 +102,7 @@ export class DashboardReaderComponent implements OnInit {
 
   ngAfterViewInit() {
     this.barecodeScanner.start();
+
   }
 
   onValueChanges(result){
